# speedup TriMesh.edgepaths by a factor 20