Based on what I did for https://github.com/python-adaptive/adaptive/commit/9f481562b0118199d4011667e857ed8eb764a2a5 I made these changes here too.

The numbers say everything :)
```python
# with this PR
%timeit trimesh.edgepaths 
4.7 ms ± 271 µs per loop (mean ± std. dev. of 7 runs, 100 loops each)

# @ master
%timeit trimesh.edgepaths
86.6 ms ± 5.65 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
```