@philippjfr 2018-02-16
There's a few other places we can use this trick. I'd be happy to work on those but I'll list the ones I can think of immediately:

[Datashader path concatenation](https://github.com/ioam/holoviews/blob/master/holoviews/operation/datashader.py#L250)

[Chord path concatenation](https://github.com/ioam/holoviews/blob/master/holoviews/element/graphs.py#L621)