@philippjfr 2018-02-16
Looks good and tests are passing, merging.